frontend/requirements.txt
# Core ML/NLP Libraries
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
spacy>=3.6.0
numpy>=1.24.0
pandas>=2.0.0

# Document Processing
PyPDF2>=3.0.1
python-docx>=0.8.11
pdfplumber>=0.9.0
pytesseract>=0.3.10

# Vector Search and Similarity
faiss-cpu>=1.7.4
chromadb>=0.4.0

# Text Processing and Preprocessing
nltk>=3.8
regex>=2023.0.0
beautifulsoup4>=4.12.0
langdetect>=1.0.9

# API and Web Framework (optional for future integration)
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0

# Utilities and Configuration
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.1.0
tqdm>=4.65.0
loguru>=0.7.0

# Data Visualization (for analysis)
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Development and Testing
pytest>=7.4.0
jupyter>=1.0.0
ipykernel>=6.24.0
